TST: refactored and moved tests related to WeightedEvaluator
